fix(memory): reconcile product and research DSA planners
