+++ CMakeLists.txt
@@ -157,6 +157,10 @@ set(PYPTO_SOURCES
     src/ir/arith/int_set.cpp
 
     # IR - Transforms
+    src/ir/transforms/dsa/allocation_plan.cpp
+    src/ir/transforms/dsa/dsa_reuse_penalty_solver.cpp
+    src/ir/transforms/dsa/memref_dsa_adapter.cpp
+    src/ir/transforms/dsa/reuse_penalty_recognizer.cpp
     src/ir/transforms/allocate_memory_addr_pass.cpp
     src/ir/transforms/canonicalize_tile_slice_pass.cpp
     src/ir/transforms/fold_no_op_reshape_pass.cpp
@@ -318,8 +322,8 @@ set(PYPTO_SOURCES
 
 if(PYPTO_ENABLE_DSA_SOLVER)
     list(APPEND PYPTO_SOURCES
-        src/ir/transforms/dsa/memref_dsa_adapter.cpp
-        src/ir/transforms/dsa/reuse_penalty_recognizer.cpp)
+        src/ir/transforms/dsa/research_memref_dsa_adapter.cpp
+        src/ir/transforms/dsa/research_reuse_penalty_recognizer.cpp)
 endif()
 
 # Include Python bindings
