+++ include/pypto/ir/transforms/dsa/reuse_penalty_recognizer.h
@@ -14,173 +14,46 @@
 
 #include <cstddef>
 #include <cstdint>
-#include <limits>
-#include <string>
 #include <vector>
 
 #include "pypto/ir/function.h"
-#include "pypto/ir/memory_space.h"
-#include "pypto/ir/transforms/pass_context.h"
-#include "pypto/ir/transforms/utils/lifetime_analysis.h"
+#include "pypto/ir/transforms/dsa/allocation_plan.h"
 
 namespace pypto {
+namespace backend {
+class Backend;
+}
 namespace ir {
 namespace dsa_adapter {
 
-enum class RecognizedReuseHazard : uint8_t {
-  SameResource,
-  CrossResource,
-};
-
-enum class RecognizedMemoryClass : uint8_t {
-  External,
-  Ub,
-  L1,
-  L0,
-  Scalar,
-};
-
-enum class RecognizedAccessResource : uint8_t {
-  InboundDma,
-  OutboundDma,
-  L0ToExternal,
-  L1ToL0,
-  L0ToL1,
-  UbToL1,
-  L1ToUb,
-  UbToL0,
-  L0ToUb,
-  VectorCompute,
-  MatrixCompute,
-  ScalarAccess,
-};
-
-struct RecognizedAccessRoute {
-  RecognizedMemoryClass source = RecognizedMemoryClass::Scalar;
-  RecognizedMemoryClass destination = RecognizedMemoryClass::Scalar;
-  RecognizedAccessResource resource = RecognizedAccessResource::ScalarAccess;
-};
-
-struct RecognizedDependencyNode {
-  size_t region = 0;
-  size_t statement_index = 0;
-};
-
-[[nodiscard]] std::string RecognizedMemoryClassToString(RecognizedMemoryClass memory_class);
-[[nodiscard]] std::string RecognizedAccessResourceToString(RecognizedAccessResource resource);
-[[nodiscard]] std::string RecognizedAccessRouteToString(const RecognizedAccessRoute& route);
-
-enum class RecognizedReuseDependence : uint8_t {
-  WriteAfterRead,
-  WriteAfterWrite,
-};
-
-struct RecognizedReuseCandidate {
-  size_t first_interval;
-  size_t second_interval;
-  size_t prior_interval;
-  size_t next_interval;
-  RecognizedReuseHazard hazard;
-  RecognizedReuseDependence dependence;
-  RecognizedAccessRoute prior_route;
-  RecognizedAccessRoute next_route;
-  MemorySpace prior_memory_space = MemorySpace::ScalarLocal;
-  MemorySpace next_memory_space = MemorySpace::ScalarLocal;
-  size_t prior_access_order = 0;
-  size_t next_access_order = 0;
-  uint64_t prior_byte_offset = 0;
-  uint64_t prior_byte_size = 0;
-  uint64_t next_byte_offset = 0;
-  uint64_t next_byte_size = 0;
-  size_t loop_id = std::numeric_limits<size_t>::max();
-  bool ordered_by_logical_dag = false;
-  std::vector<RecognizedDependencyNode> logical_order_witness;
-  bool requires_alias_contract = false;
-  bool partial_access = false;
-  bool incomplete_access_set = false;
-  bool conservative_initial_anchor = false;
-  bool nested_control = false;
-  bool in_loop = false;
-  bool loop_carried = false;
-};
-
-struct RecognizedReuseEdge {
-  size_t first_interval;
-  size_t second_interval;
-  RecognizedReuseHazard hazard;
-  bool nested_control = false;
-};
-
+/**
+ * @brief A legal but potentially costly physical-overlap relation.
+ *
+ * The interval indices refer to AllocationPlan::intervals. Cost is an abstract,
+ * non-negative optimization priority; it is not a cycle estimate.
+ */
 struct RecognizedReusePenalty {
   size_t first_interval;
   size_t second_interval;
   uint64_t cost;
-  RecognizedReuseHazard hazard;
-};
-
-struct ReusePenaltyRecognition {
-  std::vector<RecognizedReuseCandidate> candidates;
-  std::vector<RecognizedReuseEdge> edges;
-  std::vector<RecognizedReusePenalty> penalties;
-  std::vector<RecognizedAccessRoute> observed_routes;
-  size_t supported_allocations = 0;
-  size_t candidate_pairs = 0;
-  size_t already_ordered_pairs = 0;
-  size_t partially_supported_allocations = 0;
-  size_t same_resource_candidates = 0;
-  size_t cross_resource_candidates = 0;
-  size_t write_after_read_candidates = 0;
-  size_t write_after_write_candidates = 0;
-  size_t ordered_evidence_candidates = 0;
-  size_t alias_contract_candidates = 0;
-  size_t partial_access_candidates = 0;
-  size_t incomplete_access_candidates = 0;
-  size_t conservative_initial_anchor_candidates = 0;
-  size_t nested_control_candidates = 0;
-  size_t in_loop_candidates = 0;
-  size_t loop_carried_candidates = 0;
 };
 
 /**
- * @brief Recognize candidate physical-reuse hazards from PyPTO access order.
+ * @brief Recognize the compiler's built-in DSA reuse-penalty policy.
  *
- * Quadratic mode is the coverage-first research reference: it scans all
- * lifetime-compatible allocation pairs and compares per-resource access
- * frontiers, including nested and loop-carried handoffs. It is intentionally
- * super-linear and must not become a default production pass.
+ * The recognizer emits one unit-weight relation per buffer pair for which
+ * physical reuse can introduce a cross-pipe WAR or WAW handoff. It requires
+ * a complete access set, full-allocation handoff endpoints, and a verified
+ * initial write. Same-pipe, partial-view, structurally ambiguous, and
+ * uncertain handoffs remain unpenalized.
+ *
+ * The active backend supplies execution-pipe classification for supported
+ * operation and direct-memory-route combinations. The recognizer does not
+ * invoke or simulate ptoas, and skips calls whose backend pipe or physical
+ * access contract is unknown.
  */
-[[nodiscard]] ReusePenaltyRecognition RecognizeReusePenaltyCandidates(const FunctionPtr& func,
-                                                                      const AllocationPlan& allocation_plan,
-                                                                      DsaReusePenaltyRecognizer recognizer);
-
-/**
- * @brief Construct the mechanically justified experimental pair-edge subset.
- *
- * Candidate recognition records mechanism evidence. This separate policy
- * constructs one edge per qualifying cross-resource buffer pair when the
- * recognizer has a complete, full-allocation handoff, a verified minimal-write
- * antichain, and no same-operation alias question. Each abstract resource is
- * modeled as one completion-ordered issue chain. SSA def-use reachability is
- * retained as provenance but does not suppress a cross-resource handoff:
- * logical ordering does not prove asynchronous completion. Distance-zero and
- * distance-one loop handoffs are eligible. Same-resource, partial-view, and
- * uncertain candidates remain report-only.
- *
- * The policy recognizes potential synchronization obligations. It does not
- * claim that an obligation extends the consumer's exposed completion frontier;
- * that distinction belongs to the weight model once sufficient target timing
- * evidence exists.
- */
-void ConstructExperimentalPairEdges(ReusePenaltyRecognition* recognition);
-
-/**
- * @brief Apply the experimental unit weight model to constructed pair edges.
- *
- * Edge construction and weight assignment are deliberately separate: changing
- * a compiler- or device-specific cost model must not change which hazards were
- * mechanically recognized.
- */
-void ApplyExperimentalUnitPenaltyWeights(ReusePenaltyRecognition* recognition);
+[[nodiscard]] std::vector<RecognizedReusePenalty> RecognizeReusePenalties(
+    const FunctionPtr& func, const AllocationPlan& allocation_plan, const backend::Backend& backend);
 
 }  // namespace dsa_adapter
 }  // namespace ir
