+++ python/pypto/pypto_core/passes.pyi
@@ -98,6 +98,7 @@ class MemoryPlanner(Enum):
 
     PYPTO = ...
     DSA = ...
+    DSA_RP = ...
     PTOAS = ...
 
 class DsaReusePenaltyRecognizer(Enum):
